Run `make -C` instead of `cd $dir && make`
Also only enter numerical directories rather than just 4-character ones, as
there could be a 4-character directory that's not a year and maybe we don't
want to build it.

Add a collapsible section for each year and each output file/target (problems
and solutions), so that it's easy to see where the failure is, if any.